requirements.txt
langchain-google-genai
langchain-community
langchain-text-splitters
faiss-cpu
python-dotenv
pydantic